screener_project/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pandas==2.2.0
python-dateutil==2.8.2 